Factor out cadvisor
